LayoutTests/imported/w3c/web-platform-tests/inert/inert-inlines-around-selection-range-in-contenteditable-expected.txt

df

FAIL <span inert>a[bc</span><span>de]f</span> assert_equals: execCommand("delete") at <span inert>a[bc</span><span>de]f</span>: <span> content should not be deleted because anchor node of Selection is in the <span inert> expected "<span inert=\"\">abc</span><span>def</span>" but got "<span inert=\"\">abc</span><span>df</span>"
PASS <span>a[bc</span><span inert>de]f</span>
PASS <span inert>d[ef</span><span>ab]c</span>
FAIL <span>d[ef</span><span inert>ab]c</span> assert_equals: execCommand("delete") at <span>d[ef</span><span inert>ab]c</span> (selection range direction is inverted): <span> content should not be deleted because anchor node of Selection is in the <span inert> expected "<span>def</span><span inert=\"\">abc</span>" but got "<span>ef</span><span inert=\"\">abc</span>"
PASS a[bc<span inert>XYZ</span>de]f
FAIL <span inert style="display:contents">a[bc</span><span>de]f</span> assert_equals: execCommand("delete") at <span inert style="display:contents">a[bc</span><span>de]f</span>: <span> content should not be deleted because anchor node of Selection is in the <span inert> expected "<span inert=\"\" style=\"display:contents\">abc</span><span>def</span>" but got "<span inert=\"\" style=\"display:contents\">abc</span><span>df</span>"
FAIL <span inert style="display:contents">{abc</span><span>de]f</span> assert_equals: execCommand("delete") at <span inert style="display:contents">{abc</span><span>de]f</span>: <span> content should not be deleted because anchor node of Selection is in the <span inert> expected "<span inert=\"\" style=\"display:contents\">abc</span><span>def</span>" but got "<span inert=\"\" style=\"display:contents\">abc</span><span>df</span>"
FAIL <span inert><span style="display:contents">a[bc</span></span><span>de]f</span> assert_equals: execCommand("delete") at <span inert><span style="display:contents">a[bc</span></span><span>de]f</span>: <span> content should not be deleted because anchor node of Selection is in the <span inert> expected "<span inert=\"\"><span style=\"display:contents\">abc</span></span><span>def</span>" but got "<span inert=\"\"><span style=\"display:contents\">abc</span></span><span>df</span>"
FAIL <span inert><span style="display:contents">{abc</span></span><span>de]f</span> assert_equals: execCommand("delete") at <span inert><span style="display:contents">{abc</span></span><span>de]f</span>: <span> content should not be deleted because anchor node of Selection is in the <span inert> expected "<span inert=\"\"><span style=\"display:contents\">abc</span></span><span>def</span>" but got "<span inert=\"\"><span style=\"display:contents\">abc</span></span><span>df</span>"
FAIL <span inert style="display:none">a[bc</span><span>de]f</span> assert_equals: execCommand("delete") at <span inert style="display:none">a[bc</span><span>de]f</span>: <span> content should not be deleted because anchor node of Selection is in the <span inert> expected "<span inert=\"\" style=\"display:none\">abc</span><span>def</span>" but got "<span inert=\"\" style=\"display:none\">abc</span><span>df</span>"
FAIL <span inert style="display:none">{abc</span><span>de]f</span> assert_equals: execCommand("delete") at <span inert style="display:none">{abc</span><span>de]f</span>: <span> content should not be deleted because anchor node of Selection is in the <span inert> expected "<span inert=\"\" style=\"display:none\">abc</span><span>def</span>" but got "<span inert=\"\" style=\"display:none\">abc</span><span>df</span>"
FAIL <span inert><span style="display:none">a[bc</span></span><span>de]f</span> assert_equals: execCommand("delete") at <span inert><span style="display:none">a[bc</span></span><span>de]f</span>: <span> content should not be deleted because anchor node of Selection is in the <span inert> expected "<span inert=\"\"><span style=\"display:none\">abc</span></span><span>def</span>" but got "<span inert=\"\"><span style=\"display:none\">abc</span></span><span>df</span>"
FAIL <span inert><span style="display:none">{abc</span></span><span>de]f</span> assert_equals: execCommand("delete") at <span inert><span style="display:none">{abc</span></span><span>de]f</span>: <span> content should not be deleted because anchor node of Selection is in the <span inert> expected "<span inert=\"\"><span style=\"display:none\">abc</span></span><span>def</span>" but got "<span inert=\"\"><span style=\"display:none\">abc</span></span><span>df</span>"
